+++ asm/asm.go
@@ -1,0 +1,75 @@
+package asm
+
+import (
+	"fmt"
+	"log"
+
+	"github.com/kijimaD/gogo/ast"
+)
+
+func emitIntexpr(e ast.Node) {
+	switch ast := e.(type) {
+	case *ast.IntegerLiteral:
+		fmt.Printf("mov $%s, %%eax\n\t", ast.String())
+	case *ast.InfixExpression:
+		emitBinop(*ast)
+	default:
+		log.Fatal("not cover type:")
+	}
+}
+
+func emitString(e ast.StringLiteral) {
+	fmt.Printf("\t.data\n")
+	fmt.Printf(".mydata:\n\t")
+	fmt.Printf(".string \"%s\"\n\t", e.String())
+	fmt.Printf(".text\n\t")
+	fmt.Printf(".global stringfn\n")
+	fmt.Printf("stringfn:\n\t")
+	fmt.Printf("lea .mydata(%%rip), %%rax\n\t")
+	fmt.Printf("ret\n")
+}
+
+func emitBinop(i ast.InfixExpression) {
+	var op string
+	switch i.Operator {
+	case "+":
+		op = "add"
+	case "-":
+		op = "sub"
+	case "*":
+		op = "imul"
+	case "/":
+		op = ""
+	default:
+		log.Fatal("invalid operand:", op)
+	}
+
+	if i.Operator == "/" {
+		emitIntexpr(i.Left)
+		fmt.Printf("push %%rax\n\t")
+		emitIntexpr(i.Right)
+		fmt.Printf("mov %%eax, %%ebx\n\t")
+		fmt.Printf("pop %%rax\n\t")
+		fmt.Printf("mov $0, %%edx\n\t")
+		fmt.Printf("idiv %%ebx\n\t")
+	} else {
+		emitIntexpr(i.Right)
+		fmt.Printf("push %%rax\n\t")
+		emitIntexpr(i.Left)
+		fmt.Printf("pop %%rbx\n\t")
+		fmt.Printf("%s %%ebx, %%eax\n\t", op)
+	}
+}
+
+func Compile(n ast.Node) {
+	switch node := n.(type) {
+	case *ast.StringLiteral:
+		emitString(*node)
+	case *ast.InfixExpression, *ast.IntegerLiteral:
+		fmt.Printf(".text\n\t")
+		fmt.Printf(".global intfn\n")
+		fmt.Printf("intfn:\n\t")
+		emitIntexpr(node)
+		fmt.Printf("ret\n")
+	}
+}
